feat(tocco-ui): handle implicit model for tql code editor
- split code editor files up for better readability

Refs: TOCDEV-4590
Changelog: Code editors used to write TQL can now be configured in the backend to assume some entity model as the given base, allowing only writing the 'where'- and 'order by'-parts of a query.
