+++ packages/tocco-ui/src/CodeEditor/TqlConstants.js
@@ -1,0 +1,45 @@
+export const operators = ['==', '!=', '<=', '>=', '<', '>', '~=']
+export const atoms = ['true', 'false', 'null']
+export const types = [
+  'string',
+  'long',
+  'serial',
+  'bool',
+  'boolean',
+  'double',
+  'float',
+  'time',
+  'timestamp',
+  'char',
+  'varchar',
+  'date',
+  'datetime'
+]
+export const placeholders = [
+  'now_date',
+  'yesterday_date',
+  'tomorrow_date',
+  'today_date',
+  'now_start',
+  'yesterday_start',
+  'tomorrow_start',
+  'today_start',
+  'now',
+  'yesterday',
+  'tomorrow',
+  'today',
+  'currentUsername',
+  'currentUser',
+  'currentBu'
+]
+export const functions = [
+  'in',
+  'dateadd',
+  'datetimeadd',
+  'duedate',
+  'birthdayin',
+  'fulltext',
+  'likeall',
+  'distance',
+  'inboundingbox'
+]
